# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bulk SMS Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /devices:
    get:
      operationId: listDevices
      tags: [devices]
      summary: List all registered devices
      responses:
        "200":
          description: List of devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
    post:
      operationId: createDevice
      tags: [devices]
      summary: Register a new device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceInput"
      responses:
        "201":
          description: Device registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /devices/{id}:
    get:
      operationId: getDevice
      tags: [devices]
      summary: Get device by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          description: Not found
    delete:
      operationId: deleteDevice
      tags: [devices]
      summary: Remove a device
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /devices/{id}/heartbeat:
    patch:
      operationId: deviceHeartbeat
      tags: [devices]
      summary: Update device status (called from mobile)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceHeartbeat"
      responses:
        "200":
          description: Updated device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /devices/{id}/connect:
    get:
      operationId: getDeviceConnect
      tags: [devices]
      summary: Get connection info (QR code data + token)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Connection info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceConnectInfo"

  # ── Contacts ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Device:
      type: object
      required: [id, name, phoneNumber, status, token, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phoneNumber:
          type: string
        status:
          type: string
          enum: [online, offline, idle]
        batteryLevel:
          type: ["integer", "null"]
        signalStrength:
          type: ["integer", "null"]
        lastSeen:
          type: ["string", "null"]
        token:
          type: string
        createdAt:
          type: string
    DeviceInput:
      type: object
      required: [name, phoneNumber]
      properties:
        name:
          type: string
        phoneNumber:
          type: string
    DeviceHeartbeat:
      type: object
      properties:
        batteryLevel:
          type: ["integer", "null"]
        signalStrength:
          type: ["integer", "null"]
        status:
          type: ["string", "null"]
          enum: [online, offline, idle, null]
    DeviceConnectInfo:
      type: object
      required: [token, connectUrl, qrData]
      properties:
        token:
          type: string
        connectUrl:
          type: string
        qrData:
          type: string

    # ── Contact ───────────────────────────────────────────────────────────────
...