# lib/api-zod/src/generated/types/deviceHeartbeat.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Bulk SMS Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { DeviceHeartbeatStatus } from './deviceHeartbeatStatus';

export interface DeviceHeartbeat {
  /** @nullable */
  batteryLevel?: number | null;
  /** @nullable */
  signalStrength?: number | null;
  /** @nullable */
  status?: DeviceHeartbeatStatus;
}